Create Error types

+++ SysTk.WebAPI/GraphQL/Mutation.cs
@@ -1,15 +1,19 @@
 ﻿using HotChocolate.AspNetCore.Authorization;
 using HotChocolate.Execution;
 using Microsoft.AspNetCore.Identity;
+using System.ComponentModel.DataAnnotations;
 using SysTk.WebApi.Data.DataAccess;
+using Microsoft.EntityFrameworkCore;
 using SysTk.WebApi.Data.Models;
 using SysTk.WebApi.Data.Models.Auth;
 using SysTk.WebAPI.GraphQL.Errors;
+using SysTk.WebAPI.GraphQL.Types.DebugParameters;
 using SysTk.WebAPI.GraphQL.Types.DebugProcesses;
 using SysTk.WebAPI.GraphQL.Types.FtpCredential;
 using SysTk.WebAPI.GraphQL.Types.Stations;
 using SysTk.WebAPI.GraphQL.Types.Users;
 using SysTk.WebAPI.Services;
+using SysTk.WebApi.Data.Extensions;
 
 namespace SysTk.WebAPI.GraphQL
 {
@@ -18,13 +22,21 @@ namespace SysTk.WebAPI.GraphQL
         [UseDbContext(typeof(AppDbContext))]
         [Authorize(Policy = Policies.CanAdd)]
         [UseMutationConvention]
+        [Error(typeof(StationExistsError))]
         public async Task<Station> AddStationAsync(AddStationInput input,
             [ScopedService] AppDbContext context)
         {
-            // TODO: Use Automapper
+            var existingStation = context.Stations
+                .Where(x => x.IP == input.Ip || x.Id == input.Id)
+                .Select(x => new { IP = x.IP, Id = x.Id })
+                .FirstOrDefault();
+
+            if (existingStation is not null)
+                throw new StationExistsError(input.Id, input.Ip, existingStation.IP, existingStation.Id);
+            
             var station = new Station
             {
-                Id = input.Id,
+                Id = input.Id.ToUpper(),
                 Name = input.Name,
                 IP = input.Ip,
                 Cluster = input.Cluster
@@ -39,24 +51,24 @@ namespace SysTk.WebAPI.GraphQL
         [UseDbContext(typeof(AppDbContext))]
         [Authorize(Policy = Policies.CanModify)]
         [UseMutationConvention]
+        [Error(typeof(StationNotExistsError))]
         public async Task<Station> UpdateStationAsync(UpdateStationInput input,
             [ScopedService] AppDbContext context)
         {
-            var station = context.Stations.FirstOrDefault(x => x.Id == input.Id);
+            if (!context.Stations.Exists(id: input.Id))
+                throw new StationNotExistsError(input.Id);
 
-            if (station is not null)
-            {
-                station.Name = input.Name ?? station.Name;
-                station.IP = input.Ip ?? station.IP;
-                station.Cluster = input.Cluster ?? station.Cluster;
+            var station = context.Stations
+                .FirstOrDefault(x => x.Id == input.Id);
 
-                context.Stations.Update(station);
-                await context.SaveChangesAsync();
-            }
-            else
-            {
-                // TODO: Throw appropriately
-            }
+            station.Name = input.Name ?? station.Name;
+            station.IP = input.Ip ?? station.IP;
+            station.Cluster = input.Cluster ?? station.Cluster;
+
+            context.Stations.Update(station);
+            await context.SaveChangesAsync();
+
+            station.FtpCredentials = context.GetChildren(station);
 
             return station;
         }
@@ -64,11 +76,16 @@ namespace SysTk.WebAPI.GraphQL
         [UseDbContext(typeof(AppDbContext))]
         [Authorize(Policy = Policies.CanDelete)]
         [UseMutationConvention]
+        [Error(typeof(StationNotExistsError))]
         public async Task<Station> DeleteStationAsync(DeleteStationInput input,
             [ScopedService] AppDbContext context)
         {
-            // TODO: Handle station not existing
+            if (!context.Stations.Exists(id: input.Id))
+                throw new StationNotExistsError(input.Id);
+
             var station = context.Stations.Where(x => x.Id == input.Id).FirstOrDefault();
+
+            station.FtpCredentials = context.GetChildren(station);
 
             context.Stations.Attach(station);
             context.Stations.Remove(station);
@@ -80,14 +97,13 @@ namespace SysTk.WebAPI.GraphQL
         [UseDbContext(typeof(AppDbContext))]
         [Authorize(Policy = Policies.CanAdd)]
         [UseMutationConvention]
+        [Error(typeof(StationNotExistsError))]
         public async Task<FtpCredentials> AddFtpCredentialsAsync(AddFtpCredentialsInput input,
             [ScopedService] AppDbContext context)
         {
-            var stationExists = context.Stations.Where(x => x.Id == input.StationId).Any();
-            if (!stationExists)
-                throw new QueryException($"Station with ID {input.StationId} does not exist.");
+            if (!context.Stations.Exists(id: input.StationId))
+                throw new StationNotExistsError(input.StationId);
 
-            // TODO: Use AutoMapper
             var creds = new FtpCredentials
             {
                 Username = input.Username,
@@ -98,30 +114,37 @@ namespace SysTk.WebAPI.GraphQL
             context.FtpCredentials.Add(creds);
             await context.SaveChangesAsync();
 
+            creds.Station = context.GetParent(creds);
+
             return creds;
         }
 
         [UseDbContext(typeof(AppDbContext))]
         [Authorize(Policy = Policies.CanModify)]
         [UseMutationConvention]
+        [Error(typeof(StationNotExistsError))]
+        [Error(typeof(FtpCredentialsNotExistError))]
         public async Task<FtpCredentials> UpdateFtpCredentialsAsync(UpdateFtpCredentialsInput input,
             [ScopedService] AppDbContext context)
         {
-            var credentials = context.FtpCredentials.Where(x => x.StationId == input.StationId).FirstOrDefault(x => x.Username == input.Username);
+            if (!context.Stations.Exists(id: input.StationId))
+                throw new StationNotExistsError(input.StationId);
 
-            if (credentials is not null)
-            {
-                credentials.Username = input.Username ?? credentials.Username;
-                credentials.Password = input.Password ?? credentials.Password;
+            var statCreds = context.Stations.GetStationWithCredentials(input.StationId, input.Username);
 
-                context.FtpCredentials.Update(credentials);
+            if (statCreds is null)
+                throw new FtpCredentialsNotExistError(input.Username, input.StationId);
 
-                await context.SaveChangesAsync();
-            }
-            else
-            {
-                // TODO: Throw appropriately
-            }
+            var credentials = statCreds.FtpCredentials.FirstOrDefault();
+
+            credentials.Username = input.Username ?? credentials.Username;
+            credentials.Password = input.Password ?? credentials.Password;
+
+            context.FtpCredentials.Update(credentials);
+
+            await context.SaveChangesAsync();
+
+            credentials.Station = statCreds;
 
             return credentials;
         }
@@ -129,13 +152,22 @@ namespace SysTk.WebAPI.GraphQL
         [UseDbContext(typeof(AppDbContext))]
         [Authorize(Policy = Policies.CanDelete)]
         [UseMutationConvention]
+        [Error(typeof(StationNotExistsError))]
+        [Error(typeof(FtpCredentialsNotExistError))]
         public async Task<FtpCredentials> DeleteFtpCredentialsAsync(DeleteFtpCredentialsInput input,
             [ScopedService] AppDbContext context)
         {
-            // TODO: Handle credentials IP not existing
-            var credentials = context.FtpCredentials.Where(x => x.Id == input.Id ||
-            (x.StationId == input.StationId && x.Username == input.Username))
-                .FirstOrDefault();
+            if (!context.Stations.Exists(id: input.StationId))
+                throw new StationNotExistsError(input.StationId);
+
+            var statCreds = context.Stations.GetStationWithCredentials(input.StationId, input.Username, input.Id);
+
+            if (statCreds is null)
+                throw new FtpCredentialsNotExistError(input.Username, input.StationId);
+
+            var credentials = statCreds.FtpCredentials.First();
+
+            credentials.Station = statCreds;
 
             context.FtpCredentials.Attach(credentials);
             context.FtpCredentials.Remove(credentials);
@@ -147,12 +179,12 @@ namespace SysTk.WebAPI.GraphQL
         [UseDbContext(typeof(AppDbContext))]
         [Authorize(Policy = Policies.CanAdd)]
         [UseMutationConvention]
+        [Error(typeof(DebugProcessExistsError))]
         public async Task<DebugProcess> AddDebugProcessAsync(AddDebugProcessInput input,
             [ScopedService] AppDbContext context)
         {
-            var processExists = context.DebugProcesses.Where(x => x.Name == input.Name).Any();
-            if (processExists)
-                throw new QueryException($"Process with name {input.Name} already exists");
+            if (context.DebugProcesses.Exists(input.Name))
+                throw new DebugProcessExistsError(input.Name);
 
             var proc = new DebugProcess
             {
@@ -164,6 +196,142 @@ namespace SysTk.WebAPI.GraphQL
             await context.SaveChangesAsync();
 
             return proc;
+        }
+
+        [UseDbContext(typeof(AppDbContext))]
+        [Authorize(Policy = Policies.CanModify)]
+        [UseMutationConvention]
+        [Error(typeof(DebugProcessNotExistsError))]
+        public async Task<DebugProcess> UpdateDebugProcessAsync(UpdateDebugProcessInput input,
+            [ScopedService] AppDbContext context)
+        {
+            if (!context.DebugProcesses.Exists(input.Name))
+                throw new DebugProcessNotExistsError(input.Name);
+
+            var process = context.DebugProcesses
+                .First(x => x.Name.ToUpper() == input.Name.ToUpper());
+
+            process.Name = input.Name ?? process.Name;
+            process.Description = input.Description ?? process.Description;
+
+            context.DebugProcesses.Update(process);
+
+            await context.SaveChangesAsync();
+
+            process.Parameters = context.GetChildren(process);
+
+            return process;
+        }
+
+        [UseDbContext(typeof(AppDbContext))]
+        [Authorize(Policy = Policies.CanDelete)]
+        [UseMutationConvention]
+        [Error(typeof(DebugProcessNotExistsError))]
+        public async Task<DebugProcess> DeleteDebugProcessAsync(DeleteDebugProcessInput input,
+            [ScopedService] AppDbContext context)
+        {
+            if (!context.DebugProcesses.Exists(input.Id))
+                throw new DebugProcessNotExistsError(input.Id);
+
+            var process = context.DebugProcesses
+                .First(x => x.Id == input.Id);
+
+            process.Parameters = context.GetChildren(process);
+
+            context.DebugProcesses.Attach(process);
+            context.DebugProcesses.Remove(process);
+
+            await context.SaveChangesAsync();
+
+            return process;
+        }
+
+        [UseDbContext(typeof(AppDbContext))]
+        [Authorize(Policy = Policies.CanAdd)]
+        [UseMutationConvention]
+        [Error(typeof(DebugParamExistsError))]
+        public async Task<DebugParameter> AddDebugParameterAsync(AddDebugParameterInput input,
+            [ScopedService] AppDbContext context)
+        {
+            if (context.DebugParameters.Exists(input.DebugProcessId, input.Name))
+                throw new DebugParamExistsError(input.Name, input.DebugProcessId);
+
+            var param = new DebugParameter
+            {
+                Name = input.Name,
+                Description = input.Description,
+                DebugProcessId = input.DebugProcessId
+            };
+
+            context.DebugParameters.Add(param);
+            await context.SaveChangesAsync();
+
+            param = context.DebugParameters
+                .FirstOrDefault(x => x.Name == input.Name);
+
+            param.Process = context.GetParent(param);
+
+            return param;
+        }
+
+        [UseDbContext(typeof(AppDbContext))]
+        [Authorize(Policy = Policies.CanModify)]
+        [UseMutationConvention]
+        [Error(typeof(DebugProcessNotExistsError))]
+        [Error(typeof(DebugParameterNotExistsError))]
+        public async Task<DebugParameter> UpdateDebugParameterAsync(UpdateDebugParameterInput input,
+            [ScopedService] AppDbContext context)
+        {
+            if (!context.DebugProcesses.Exists(input.DebugProcessName))
+                throw new DebugProcessNotExistsError(input.DebugProcessName);
+
+            var process = context.DebugProcesses.GetProcessWithParameters(input.DebugProcessName, input.Name);
+
+            if (process is null)
+                throw new DebugParameterNotExistsError(input.Name, input.DebugProcessName);
+
+            var param = process.Parameters.First();
+
+            param.Process = process;
+
+            param.Name = input.Name ?? param.Name;
+            param.Description = input.Description ?? param.Description;
+
+            context.DebugParameters.Update(param);
+
+            await context.SaveChangesAsync();
+
+            return param;
+        }
+
+        [UseDbContext(typeof(AppDbContext))]
+        [Authorize(Policy = Policies.CanDelete)]
+        [UseMutationConvention]
+        [Error(typeof(DebugProcessNotExistsError))]
+        [Error(typeof(DebugParameterNotExistsError))]
+        public async Task<DebugParameter> DeleteDebugParameterAsync(DeleteDebugParameterInput input,
+            [ScopedService] AppDbContext context)
+        {
+            if (!context.DebugProcesses.Exists(input.DebugProcessName))
+                throw new DebugProcessNotExistsError(input.Name);
+
+            var process = context.DebugProcesses.GetProcessWithParameters(input.DebugProcessName, input.Name);
+
+            if (process is null)
+                throw new DebugParameterNotExistsError(input.Name, input.DebugProcessName);
+
+            var param = process.Parameters.First();
+
+            var paramProcess = context.GetParent(param);
+
+            param.Process = paramProcess;
+
+            context.DebugParameters.Attach(param);
+            context.DebugParameters.Remove(param);
+
+            await context.SaveChangesAsync();
+
+            return param;
         }
 
         [UseDbContext(typeof(AppDbContext))]
@@ -183,9 +351,7 @@ namespace SysTk.WebAPI.GraphQL
             var userCreateResult = await userManager.CreateAsync(user, input.Password);
 
             if (userCreateResult.Succeeded)
-            {
                 return user;
-            }
 
             throw new QueryException($"User could not be added: {userCreateResult.Errors}");
         }
@@ -225,18 +391,15 @@ namespace SysTk.WebAPI.GraphQL
             throw new QueryException("Unknown error.");
         }
 
+        [Error(typeof(LoginFailedError))]
         public async Task<LoginPayload> Login(LoginInput input,
             [Service] ITokenService tokenService)
         {
-            if (await tokenService.IsValidUsernameAndPassword(input.Username, input.Password))
-            {
-                var output = await tokenService.GenerateToken(input.Username);
-                return new LoginPayload { AccessToken = output.AccessToken, Username = output.Username };
-            }
-            else
-            {
-                throw new QueryException(ErrorFactory.CreateLoginError());
-            }
+            if (!await tokenService.IsValidUsernameAndPassword(input.Username, input.Password))
+                throw new LoginFailedError();
+
+            var output = await tokenService.GenerateToken(input.Username);
+            return new LoginPayload(output.AccessToken, output.Username);  
         }
     }
 }
